Make generated rust async futures Send
The futures for generated rust async bindings were not always Send,
because the LoweredParams held a non-Send `*mut u8`. To fix this,
explicitly implement Send for LoweredParams. This makes the futures
compatible with more of the Rust ecosystem, eg. an axum webserver where
all handlers must be Send. This is safe, at least for now, because
the generated code is single threaded.

To make LoweredParams Send, change the type from a tuple to a struct
which we can explicitly mark as Send. This also requires moving it up a
little bit in the generated code outside of the trait impl.

Add a test that asserts the generated binding is Send.